vlm-core/requirements.txt
fastapi==0.103.1
pydantic==2.3.0
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.0
numpy==1.25.2
requests==2.31.0
httpx==0.25.0
opencv-python==4.8.0.76
langchain==0.0.267
langchain-core==0.1.12
transformers==4.37.2
accelerate==0.28.0
torch==2.1.2
sentencepiece==0.1.99
